Add error handling to settings and fix types

+++ modpack_builder/gui/settings.py
@@ -4,6 +4,7 @@ import shutil
 import platform
 
 from pathlib import Path
+from json import JSONDecodeError
 
 from . import PROGRAM_NAME
 
@@ -19,6 +20,10 @@ class ModpackBuilderSettings:
 
         self.builder = builder
 
+        self.__settings_directory = None
+        self.__settings_file = None
+        self.__curseforge_cache_file = None
+
         if path:
             settings_directory = path
         else:
@@ -32,27 +37,38 @@ class ModpackBuilderSettings:
         self.curseforge_cache = dict()
 
     def load_settings(self):
-        if not self.__settings_file.exists() and not self.__settings_file.is_file():
+        if not self.__settings_file.exists() or not self.__settings_file.is_file():
             return False
 
-        with open(self.__settings_file, "r") as file:
-            data = json.load(file)
+        try:
+            with open(self.__settings_file, "r") as file:
+                data = json.load(file)
 
-            self.builder.concurrent_requests = data.get("concurrent_requests", self.builder.concurrent_requests)
-            self.builder.concurrent_downloads = data.get("concurrent_downloads", self.builder.concurrent_downloads)
-            self.builder.minecraft_directory = data.get("minecraft_directory", self.builder.minecraft_directory)
-            self.builder.minecraft_launcher_path = data.get("minecraft_launcher_path",
-                                                            self.builder.minecraft_launcher_path)
-            self.builder.profiles_directory = data.get("profiles_directory", self.builder.profiles_directory)
+                self.builder.concurrent_requests = data.get("concurrent_requests", self.builder.concurrent_requests)
+                self.builder.concurrent_downloads = data.get("concurrent_downloads", self.builder.concurrent_downloads)
+                self.builder.minecraft_directory = Path(
+                    data.get("minecraft_directory", self.builder.minecraft_directory))
+                self.builder.minecraft_launcher_path = Path(
+                    data.get("minecraft_launcher_path", self.builder.minecraft_launcher_path))
+                self.builder.profiles_directory = Path(data.get("profiles_directory", self.builder.profiles_directory))
+
+        except JSONDecodeError:
+            self.__settings_file.unlink()
+            return False
 
         return True
 
     def load_curseforge_cache(self):
-        if not self.__curseforge_cache_file.exists() and not self.__curseforge_cache_file.is_file():
+        if not self.__curseforge_cache_file.exists() or not self.__curseforge_cache_file.is_file():
             return False
 
-        with open(self.__curseforge_cache_file, "r") as file:
-            self.curseforge_cache.update(json.load(file))
+        try:
+            with open(self.__curseforge_cache_file, "r") as file:
+                self.curseforge_cache.update(json.load(file))
+
+        except JSONDecodeError:
+            self.__curseforge_cache_file.unlink()
+            return False
 
         return True
 
@@ -67,12 +83,15 @@ class ModpackBuilderSettings:
     @staticmethod
     def get_settings_directory():
         if PLATFORM == "Windows":
-            program_key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, rf"SOFTWARE\{PROGRAM_NAME}")
-            settings_path = Path(os.path.expandvars(winreg.QueryValueEx(program_key, "SettingsDirectory")[0]))
+            try:
+                program_key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, rf"SOFTWARE\{PROGRAM_NAME}")
+                settings_path = Path(os.path.expandvars(winreg.QueryValueEx(program_key, "SettingsDirectory")[0]))
 
-            program_key.Close()
+                program_key.Close()
 
-            return settings_path
+                return settings_path
+            except FileNotFoundError:
+                return None
 
         elif PLATFORM == "Darwin":
             return None
@@ -85,7 +104,7 @@ class ModpackBuilderSettings:
         if PLATFORM == "Windows":
             program_key = winreg.CreateKeyEx(winreg.HKEY_CURRENT_USER, rf"SOFTWARE\{PROGRAM_NAME}")
 
-            winreg.SetValueEx(program_key, "SettingsDirectory", None, winreg.REG_EXPAND_SZ, path.resolve())
+            winreg.SetValueEx(program_key, "SettingsDirectory", None, winreg.REG_EXPAND_SZ, str(path.resolve()))
             winreg.FlushKey(program_key)
 
             program_key.Close()
@@ -98,16 +117,16 @@ class ModpackBuilderSettings:
 
     @property
     def settings_directory(self):
-        return self.__settings_path
+        return self.__settings_directory
 
     @settings_directory.setter
     def settings_directory(self, value):
         (value := value.resolve()).mkdir(parents=True, exist_ok=True)
 
-        if self.settings_directory.exists() and self.settings_directory.is_dir():
+        if self.settings_directory and self.settings_directory.exists() and self.settings_directory.is_dir():
             shutil.move(str(self.settings_directory), str(value))
 
-        self.__settings_path = value
+        self.__settings_directory = value
         self.__settings_file = value / "settings.json"
         self.__curseforge_cache_file = value / "curseforge_cache.json"
 
@@ -163,7 +182,7 @@ class ModpackBuilderSettings:
         return {
             "concurrent_requests": self.concurrent_requests,
             "concurrent_downloads": self.concurrent_downloads,
-            "minecraft_directory": self.minecraft_directory.resolve(),
-            "minecraft_launcher_path": self.minecraft_launcher_path.resolve(),
-            "profiles_directory": self.profiles_directory.resolve()
+            "minecraft_directory": str(self.minecraft_directory.resolve()),
+            "minecraft_launcher_path": str(self.minecraft_launcher_path.resolve()),
+            "profiles_directory": str(self.profiles_directory.resolve())
         }
